Updates name of 'router' folder to 'lambdaData' and changes nested file's name from 'router.js' to 'lambdaDataRouter.js'
